requirements_dev.txt
djangorestframework
django_filter
django-fsm
strawberry-django-plus
uwsgi
psycopg2-binary
whitenoise
flake8
mypy
autoflake
black
isort[requirements_deprecated_finder]
requests_oauthlib
PyJWT
